Finally, implementations for IBinaryWriter and IBinaryReader have arrived.

+++ Vostok.Core/Commons/Binary/IBinaryWriter.cs
@@ -5,7 +5,7 @@ namespace Vostok.Commons.Binary
 {
     public interface IBinaryWriter
     {
-        long Position { get; set; }
+        int Position { get; set; }
 
         void Write(int value);
         void Write(long value);
